Code generation for SAV Library MVC 11.5.0 modified

+++ Classes/CodeGenerator/CodeGeneratorForSavLibraryMvc.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of the TYPO3 CMS project.
  *
@@ -23,9 +25,8 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  * It is based on the same idea developed by Ingmar Schlecht for the extbase_kickstater.
  * Code templates are used to build the file contents. They are processed by a fluid parser.
  */
-class CodeGeneratorForSavLibraryMvc extends AbstractCodeGenerator
+class CodeGeneratorForSavLibraryMvc extends PreprocessingForCodeGenerator
 {
-
     /**
      * The code templates directory
      *
@@ -44,6 +45,9 @@ class CodeGeneratorForSavLibraryMvc extends AbstractCodeGenerator
         if (! $this->CanBuildExtension()) {
             return;
         }
+
+        // Preprocessing for MVC
+        $this->buildArrayForCodeGenerator(1);
 
         // Generates the Icons
         $this->buildIcons();
@@ -64,6 +68,9 @@ class CodeGeneratorForSavLibraryMvc extends AbstractCodeGenerator
         $this->buildConfigurationFlexform();
         $this->buildConfigurationTca();
         $this->buildConfigurationTypoScript();
+        $this->buildConfigurationServices();
+        $this->buildConfigurationExtbasePersistence();
+        $this->buildConfigurationRoutes();
 
         // Generates Documentation files
         $this->buildDocumentation();
@@ -79,6 +86,9 @@ class CodeGeneratorForSavLibraryMvc extends AbstractCodeGenerator
 
         // Generates the Controller (must be the last generated part for subforms)
         $this->buildController();
+
+        // Generates the migration file if requested
+        $this->buildMigration();
     }
 
     /**
@@ -94,17 +104,61 @@ class CodeGeneratorForSavLibraryMvc extends AbstractCodeGenerator
     {
         GeneralUtility::mkdir_deep($this->extensionDirectory . 'Configuration/TypoScript/');
 
-        // Generates constants.txt file if it does not exist
-        if (! file_exists($this->extensionDirectory . 'Configuration/TypoScript/constants.txt')) {
-            $fileContents = $this->generateFile('Configuration/TypoScript/constants.txtt');
-            GeneralUtility::writeFile($this->extensionDirectory . 'Configuration/TypoScript/constants.txt', $fileContents);
+        // Generates constants.typoscript file if it does not exist
+        if (! file_exists($this->extensionDirectory . 'Configuration/TypoScript/constants.typoscript')) {
+            $fileContents = $this->generateFile('Configuration/TypoScript/constants.typoscriptt');
+            GeneralUtility::writeFile($this->extensionDirectory . 'Configuration/TypoScript/constants.typoscript', $fileContents);
         }
 
-        // Generates setup.txt file if it does not exist
-        if (! file_exists($this->extensionDirectory . 'Configuration/TypoScript/setup.txt')) {
-            $fileContents = $this->generateFile('Configuration/TypoScript/setup.txtt');
-            GeneralUtility::writeFile($this->extensionDirectory . 'Configuration/TypoScript/setup.txt', $fileContents);
+        // Generates setup.typoscript file if it does not exist
+        if (! file_exists($this->extensionDirectory . 'Configuration/TypoScript/setup.typoscript')) {
+            $fileContents = $this->generateFile('Configuration/TypoScript/setup.typoscriptt');
+            GeneralUtility::writeFile($this->extensionDirectory . 'Configuration/TypoScript/setup.typoscript', $fileContents);
         }
+    }
+
+    /**
+     * Builds the documentation.
+     *
+     * @return void
+     */
+    protected function buildDocumentation()
+    {
+        // Generates the parent documentation
+        parent::buildDocumentation();
+
+        // Removes existing directories
+        $this->removeDirectory('Documentation/SavLibraryConfiguration');
+
+        if ($this->sectionManager->getItem('documentation')
+            ->getItem(1)
+            ->getItem('AddFormAndTableConfiguration')) {
+
+                // Creates the folders
+                GeneralUtility::mkdir_deep($this->extensionDirectory . 'Documentation/SavLibraryConfiguration/Forms/');
+
+                // Prepares the data
+                $extension = $this->sectionManager->getItemsAsArray();
+                $data = $extension;
+                $data['extensionDirectory'] = $this->extensionDirectory;
+                $data['viewsFields'] = $this->arrayForCodeGenerator['views'];
+
+                // Generates the specific documentation for SAV Library Mvc
+                $fileContents = $this->generateFile('Documentation/SavLibraryConfiguration/Index.rstt', null, $data);
+                GeneralUtility::writeFile($this->extensionDirectory . 'Documentation/SavLibraryConfiguration/Index.rst', $fileContents);
+                $fileContents = $this->generateFile('Documentation/SavLibraryConfiguration/Forms/Index.rstt', null, $data);
+                GeneralUtility::writeFile($this->extensionDirectory . 'Documentation/SavLibraryConfiguration/Forms/Index.rst', $fileContents);
+                if (!empty($extension['newTables'])) {
+                    GeneralUtility::mkdir_deep($this->extensionDirectory . 'Documentation/SavLibraryConfiguration/NewTables');
+                    $fileContents = $this->generateFile('Documentation/SavLibraryConfiguration/NewTables/Index.rstt', null, $data);
+                    GeneralUtility::writeFile($this->extensionDirectory . 'Documentation/SavLibraryConfiguration/NewTables/Index.rst', $fileContents);
+                }
+                if (!empty($extension['existingTables'])) {
+                    GeneralUtility::mkdir_deep($this->extensionDirectory . 'Documentation/SavLibraryConfiguration/ExistingTables');
+                    $fileContents = $this->generateFile('Documentation/SavLibraryConfiguration/ExistingTables/Index.rstt', null, $data);
+                    GeneralUtility::writeFile($this->extensionDirectory . 'Documentation/SavLibraryConfiguration/ExistingTables/Index.rst', $fileContents);
+                }
+            }
     }
 
     /**
@@ -114,12 +168,15 @@ class CodeGeneratorForSavLibraryMvc extends AbstractCodeGenerator
      */
     protected function buildController()
     {
+        // Removes existing directories
+        $this->removeDirectory('Classes/Controller');
+
         GeneralUtility::mkdir_deep($this->extensionDirectory . 'Classes/Controller/');
         $fileDirectory = $this->extensionDirectory . 'Classes/Controller/';
-        foreach ($this->sectionManager->getItem('forms')->getItems() as $itemKey => $item) {
+        foreach ($this->sectionManager->getItem('forms') as $formKey => $form) {
             // Every form gets a corresponding Action Controller
-            $fileContents = $this->generateFile('Classes/Controller/Controller.phpt', $itemKey);
-            GeneralUtility::writeFile($fileDirectory . GeneralUtility::underscoredToUpperCamelCase($item['title']) . 'Controller.php', $fileContents);
+            $fileContents = $this->generateFile('Classes/Controller/Controller.phpt', $formKey);
+            GeneralUtility::writeFile($fileDirectory . GeneralUtility::underscoredToUpperCamelCase($form->getItem('title')) . 'Controller.php', $fileContents);
         }
     }
 
@@ -130,12 +187,24 @@ class CodeGeneratorForSavLibraryMvc extends AbstractCodeGenerator
      */
     protected function buildDomainModels()
     {
+        // Removes existing directories
+        $this->removeDirectory('Classes/Domain/Model');
+
         GeneralUtility::mkdir_deep($this->extensionDirectory . 'Classes/Domain/Model/');
         $fileDirectory = $this->extensionDirectory . 'Classes/Domain/Model/';
-        foreach ($this->sectionManager->getItem('newTables')->getItems() as $itemKey => $item) {
-            // Every table has a domain model
-            $fileContents = $this->generateFile('Classes/Domain/Model/Model.phpt', $itemKey);
-            GeneralUtility::writeFile($fileDirectory . GeneralUtility::underscoredToUpperCamelCase($item['tablename']) . '.php', $fileContents);
+        if ($this->sectionManager->getItem('newTables') !== null) {
+            foreach ($this->sectionManager->getItem('newTables')->getItems() as $itemKey => $item) {
+                // Every table has a domain model
+                $fileContents = $this->generateFile('Classes/Domain/Model/ModelForNewTables.phpt', $itemKey);
+                GeneralUtility::writeFile($fileDirectory . GeneralUtility::underscoredToUpperCamelCase($item['tablename']) . '.php', $fileContents);
+            }
+        }
+        if ($this->sectionManager->getItem('existingTables') !== null) {
+            foreach ($this->sectionManager->getItem('existingTables')->getItems() as $itemKey => $item) {
+                // Every table has a domain model
+                $fileContents = $this->generateFile('Classes/Domain/Model/ModelForExistingTables.phpt', $itemKey);
+                GeneralUtility::writeFile($fileDirectory . GeneralUtility::underscoredToUpperCamelCase($item['tablename']) . '.php', $fileContents);
+            }
         }
     }
 
@@ -146,12 +215,70 @@ class CodeGeneratorForSavLibraryMvc extends AbstractCodeGenerator
      */
     protected function buildDomainRepositories()
     {
+        // Removes existing directories
+        $this->removeDirectory('Classes/Domain/Repository');
+
         GeneralUtility::mkdir_deep($this->extensionDirectory . 'Classes/Domain/Repository/');
         $fileDirectory = $this->extensionDirectory . 'Classes/Domain/Repository/';
-        foreach ($this->sectionManager->getItem('newTables')->getItems() as $itemKey => $item) {
-            // Every table has a domain repository
-            $fileContents = $this->generateFile('Classes/Domain/Repository/Repository.phpt', $itemKey);
-            GeneralUtility::writeFile($fileDirectory . GeneralUtility::underscoredToUpperCamelCase($item['tablename']) . 'Repository.php', $fileContents);
+        if ($this->sectionManager->getItem('newTables') !== null) {
+            foreach ($this->sectionManager->getItem('newTables')->getItems() as $itemKey => $item) {
+                // Every table has a domain repository
+                $fileContents = $this->generateFile('Classes/Domain/Repository/RepositoryForNewTables.phpt', $itemKey);
+                GeneralUtility::writeFile($fileDirectory . GeneralUtility::underscoredToUpperCamelCase($item['tablename']) . 'Repository.php', $fileContents);
+            }
+        }
+        if ($this->sectionManager->getItem('existingTables') !== null) {
+            foreach ($this->sectionManager->getItem('existingTables')->getItems() as $itemKey => $item) {
+                // Every table has a domain repository
+                $fileContents = $this->generateFile('Classes/Domain/Repository/RepositoryForExistingTables.phpt', $itemKey);
+                GeneralUtility::writeFile($fileDirectory . GeneralUtility::underscoredToUpperCamelCase($item['tablename']) . 'Repository.php', $fileContents);
+            }
+        }
+    }
+
+    /**
+     * Builds the Configuration/Extbase/Persistence/Classes.php file.
+     *
+     * @return void
+     */
+    protected function buildConfigurationExtbasePersistence()
+    {
+        $fileDirectory = $this->extensionDirectory . 'Configuration/Extbase/Persistence/';
+        GeneralUtility::mkdir_deep($fileDirectory);
+        $fileContents = $this->generateFile('Configuration/Extbase/Persistence/Classes.phpt');
+        GeneralUtility::writeFile($fileDirectory . 'Classes.php', $fileContents);
+    }
+
+    /**
+     * Builds the Configuration/Routes/Default.yaml file.
+     *
+     * @return void
+     */
+    protected function buildConfigurationRoutes()
+    {
+        $fileDirectory = $this->extensionDirectory . 'Configuration/Routes/';
+        GeneralUtility::mkdir_deep($fileDirectory);
+        $fileContents = $this->generateFile('Configuration/Routes/Default.yamlt');
+        GeneralUtility::writeFile($fileDirectory . 'Default.yaml', $fileContents);
+    }
+
+
+    /**
+     * Builds the migration SQL file.
+     *
+     * @return void
+     */
+    protected function buildMigration()
+    {
+        if ($this->sectionManager->getItem('general')
+            ->getItem(1)
+            ->getItem('BuildMigration')) {
+
+            GeneralUtility::mkdir_deep($this->extensionDirectory . 'Migration/');
+            $fileDirectory = $this->extensionDirectory . 'Migration/';
+
+            $fileContents = $this->generateFile('Migration/MigrationFromSavLibraryPlus.sqlt');
+            GeneralUtility::writeFile($fileDirectory . 'MigrationFromSavLibraryPlus.sql', $fileContents);
         }
     }
 }
